♻️  Refactoring test typo
