+++ src/pages/restaurant/RestaurantPage.jsx
@@ -4,9 +4,9 @@ import { useDispatch } from 'react-redux';
 
 import { useParams } from 'react-router-dom';
 
-import { loadRestaurant } from '../actions';
+import { loadRestaurant } from '../../actions';
 
-import RestaurantContainer from '../containers/RestaurantContainer';
+import RestaurantContainer from '../../containers/RestaurantContainer';
 
 export default function RestaurantPage() {
   const dispatch = useDispatch();
