+++ src/pages/restaurant/RestaurantPage.test.jsx
@@ -5,6 +5,7 @@ import { render } from '@testing-library/react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import { MemoryRouter } from 'react-router-dom';
+
 import RestaurantPage from './RestaurantPage';
 
 describe('RestaurantPage', () => {
